move newspecobj over
