+++ pypeit/pypeit.py
@@ -14,7 +14,7 @@ from pypeit import reduce
 from pypeit.core import qa
 from pypeit.core import wave
 from pypeit.core import save
-from pypeit import newspecobjs
+from pypeit import specobjs
 from pypeit.core import pixels
 from pypeit.spectrographs.util import load_spectrograph
 from linetools import utils as ltu
@@ -485,7 +485,7 @@ class PypeIt(object):
 
         """
         if std_redux is False and std_outfile is not None:
-            sobjs = newspecobjs.SpecObjs.from_fitsfile(std_outfile)
+            sobjs = specobjs.SpecObjs.from_fitsfile(std_outfile)
             # Does the detector match?
             # TODO Instrument specific logic here could be implemented with the parset. For example LRIS-B or LRIS-R we
             # we would use the standard from another detector
@@ -673,8 +673,8 @@ class PypeIt(object):
         # Determine the paths/filenames
         save.save_all(sci_dict, self.caliBrate.master_key_dict, self.caliBrate.master_dir,
                       self.spectrograph, head1d, head2d, self.science_path, basename,
-                      update_det=self.par['rdx']['detnum'],
-                      binning=self.fitstbl['binning'][frame])
+                      update_det=self.par['rdx']['detnum'])
+                      #binning=self.fitstbl['binning'][frame])
 
     def msgs_reset(self):
         """
